Split off conda_spawn.contrib and conda_spawn.registry modules
Move best-effort shells (fish, csh/tcsh, xonsh) into a dedicated
`conda_spawn.contrib` subpackage to make their community/best-effort
status explicit at the code level, matching the README tier split and
addressing review feedback on PR #28.

Extract the `SHELLS` dispatch table, `default_shell_class`, and
`detect_shell_class` into `conda_spawn.registry`. This keeps `shell.py`
focused on shell base classes and fully-supported implementations, and
avoids a circular import between `shell.py` and `contrib.py` (the
registry depends on both; nothing depends back on it).
