+++ conda_spawn/shell.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import os
-import re
 import shlex
 import shutil
 import signal
@@ -18,8 +17,6 @@ if sys.platform != "win32":
     import termios
 
     import pexpect
-
-import shellingham
 
 from . import activate
 
@@ -242,142 +239,6 @@ class ZshShell(PosixShell):
         return "zsh"
 
 
-class FishShell(UnixShell):
-    """fish shell support.
-
-    Best-effort: not part of the core supported shell matrix (`bash`,
-    `zsh`, `powershell`, `cmd`). It exists to bring `conda spawn` to
-    parity with `conda activate`'s own activator coverage, but is
-    lightly tested and relies on user reports for shell-specific bugs.
-    """
-
-    Activator = activate.FishActivator
-    default_shell = "fish"
-
-    def prompt(self) -> str:
-        # Preserve any pre-existing `fish_prompt` (including ones installed
-        # by prompt tools like starship) and prepend conda's prompt modifier
-        # so users see their env name regardless of their shell setup.
-        return (
-            "if functions -q fish_prompt\n"
-            "    if not functions -q __conda_spawn_orig_fish_prompt\n"
-            "        functions -c fish_prompt __conda_spawn_orig_fish_prompt\n"
-            "        functions -e fish_prompt\n"
-            "    end\n"
-            "end\n"
-            "function fish_prompt\n"
-            '    printf "%s" "$CONDA_PROMPT_MODIFIER"\n'
-            "    if functions -q __conda_spawn_orig_fish_prompt\n"
-            "        __conda_spawn_orig_fish_prompt\n"
-            "    end\n"
-            "end"
-        )
-
-    def source_command(self, script_path: str) -> str:
-        return f'source "{script_path}"'
-
-    def executable(self) -> str:
-        return "fish"
-
-
-class CshShell(UnixShell):
-    """csh shell support.
-
-    Best-effort: not part of the core supported shell matrix (`bash`,
-    `zsh`, `powershell`, `cmd`). It exists to bring `conda spawn` to
-    parity with `conda activate`'s own activator coverage, but is
-    lightly tested and relies on user reports for shell-specific bugs.
-    """
-
-    Activator = activate.CshActivator
-    default_shell = "csh"
-    default_args = ("-i",)
-    prompt_strip_markers = ("set prompt=",)
-
-    def prompt(self) -> str:
-        # csh/tcsh do not define $prompt automatically in all modes; guard
-        # against "Undefined variable" by initialising it to "" if absent.
-        return (
-            'if (! $?prompt) set prompt = ""\n'
-            f'set prompt="{self.prompt_modifier()}${{prompt}}"'
-        )
-
-    def source_command(self, script_path: str) -> str:
-        return f'source "{script_path}"'
-
-    def ready_marker_command(self) -> str:
-        # csh does not ship a `printf` builtin; `echo -n` is portable
-        # across csh/tcsh on the platforms we support.
-        return f'echo -n "{self.READY_MARKER}"'
-
-    def executable(self) -> str:
-        return "csh"
-
-
-class TcshShell(CshShell):
-    """tcsh shell support.
-
-    Inherits the same best-effort caveat as `CshShell`.
-    """
-
-    default_shell = "tcsh"
-
-    def executable(self) -> str:
-        return "tcsh"
-
-
-class XonshShell(UnixShell):
-    """xonsh shell support.
-
-    Best-effort: not part of the core supported shell matrix (`bash`,
-    `zsh`, `powershell`, `cmd`). It exists to bring `conda spawn` to
-    parity with `conda activate`'s own activator coverage, but is
-    lightly tested and relies on user reports for shell-specific bugs.
-    """
-
-    Activator = activate.XonshActivator
-    default_shell = "xonsh"
-    default_args = ("-i",)
-
-    @property
-    def script_suffix(self) -> str:
-        # The `XonshActivator` reports `.sh` (for bash-sourced
-        # `activate.d` scripts) but `execute()` emits xonsh syntax.
-        # `.xsh` is the canonical extension for xonsh scripts.
-        return ".xsh"
-
-    def script(self) -> str:
-        # `XonshActivator.unset_var_tmpl` emits `del $VAR` which raises
-        # `KeyError` when the variable is not already set (e.g. on a fresh
-        # CI runner).  Replace every such line with the safe pop form.
-        raw = super().script()
-        return re.sub(
-            r"^del \$(\w+)$",
-            lambda m: f'${{...}}.pop("{m.group(1)}", None)',
-            raw,
-            flags=re.MULTILINE,
-        )
-
-    def prompt(self) -> str:
-        # Prepend `CONDA_PROMPT_MODIFIER` to `$PROMPT` while keeping
-        # any format-field tokens in the user's prompt intact.
-        return "$PROMPT = ${...}.get('CONDA_PROMPT_MODIFIER', '') + $PROMPT"
-
-    def source_command(self, script_path: str) -> str:
-        return f'source "{script_path}"'
-
-    def post_activation_command(self) -> str:
-        # In xonsh, bare `stty echo` is treated as subprocess but the
-        # explicit `$[...]` form is unambiguous inside a sourced script.
-        return "$[stty echo]"
-
-    def ready_marker_command(self) -> str:
-        return f'print({self.READY_MARKER!r}, end="", flush=True)'
-
-    def executable(self) -> str:
-        return "xonsh"
-
-
 class PowershellShell(Shell):
     Activator = activate.PowerShellActivator
 
@@ -457,39 +318,3 @@ class CmdExeShell(PowershellShell):
 
     def args(self) -> tuple[str, ...]:
         return ("/D", "/K")
-
-
-SHELLS: dict[str, type[Shell]] = {
-    "ash": PosixShell,
-    "bash": BashShell,
-    "cmd.exe": CmdExeShell,
-    "cmd": CmdExeShell,
-    "csh": CshShell,
-    "dash": PosixShell,
-    "fish": FishShell,
-    "posix": PosixShell,
-    "powershell": PowershellShell,
-    "pwsh": PowershellShell,
-    "tcsh": TcshShell,
-    "xonsh": XonshShell,
-    "zsh": ZshShell,
-}
-
-
-def default_shell_class():
-    if sys.platform == "win32":
-        return CmdExeShell
-    return PosixShell
-
-
-def detect_shell_class():
-    try:
-        name, _ = shellingham.detect_shell()
-    except shellingham.ShellDetectionFailure:
-        return default_shell_class()
-    else:
-        try:
-            return SHELLS[name]
-        except KeyError:
-            log.warning("Did not recognize shell %s, returning default.", name)
-            return default_shell_class()
